Group utils.py files in a single package
