+++ squalaetp/management/commands/corvet.py
@@ -6,7 +6,7 @@ from django.conf import settings
 
 from squalaetp.models import Corvet
 
-from ._excel_format import ExcelSqualaetp
+from ._excel_squalaetp import ExcelSqualaetp
 
 import logging as log
