+++ squalaetp/models.py
@@ -3,14 +3,14 @@ from django.db import models
 
 class Xelon(models.Model):
     numero_de_dossier = models.CharField('numéro de dossier', max_length=10, unique=True)
-    vin = models.CharField('V.I.N.', max_length=17, null=True, blank=True)
+    vin = models.CharField('V.I.N.', max_length=17, blank=True)
     modele_produit = models.CharField('modèle produit', max_length=50, blank=True)
     modele_vehicule = models.CharField('modèle véhicule', max_length=50, blank=True)
     famille_client = models.CharField('famille Client', max_length=5000, blank=True)
     famille_produit = models.CharField('famille produit', max_length=100, blank=True)
-    date_retour = models.DateField('date retour', null=True)
-    delai_au_en_jours_ouvres = models.IntegerField('délai en jours ouvrés', null=True)
-    delai_au_en_jours_calendaires = models.IntegerField('délai en jours calendaires', null=True)
+    date_retour = models.DateField('date retour', null=True, blank=True)
+    delai_au_en_jours_ouvres = models.IntegerField('délai en jours ouvrés', null=True, blank=True)
+    delai_au_en_jours_calendaires = models.IntegerField('délai en jours calendaires', null=True, blank=True)
     date_de_cloture = models.DateTimeField('date de clôture', null=True, blank=True)
     type_de_cloture = models.CharField('type de clôture', max_length=50, blank=True)
     lieu_de_stockage = models.CharField('lieu de stockage', max_length=50, blank=True)
